api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.3
python-dotenv==1.0.1
pytz==2024.2